+++ backend/prisma/schema.prisma
@@ -24,6 +24,8 @@ model User {
 
   accounts Account[]
 
+  forms Form[]
+
   isVerified Boolean @default(false) @map("is_verified")
   isTwoFactorEnabled Boolean @default(false) @map("is_two_factor_enabled")
 
@@ -78,6 +80,8 @@ enum TokenType {
 }
 
 model Form {
+  user User @relation(fields: [userId], references: [id])
+  userId String  @map("user_id")
   id        String   @id @default(uuid())
   fields    Field[]
   createdAt DateTime @default(now()) @map("created_at")
@@ -91,7 +95,7 @@ model Field {
   type   FieldType
   label  String  
   form   Form     @relation(fields: [formId], references: [id])
-  formId String  
+  formId String   @map("form_id")
 
   @@map("fields")
 }
