+++ backend/src/app.module.ts
@@ -3,11 +3,11 @@ import { ConfigModule } from '@nestjs/config'
 
 import { AppController } from './app.controller'
 import { AppService } from './app.service'
-import { AuthModule } from './auth/auth.module'
 import { IS_DEV_ENV } from './libs/common/utils/is-dev.util'
-import { PrismaModule } from './prisma/prisma.module'
-import { UserModule } from './user/user.module';
-import { FormModule } from './form/form.module';
+import { AuthModule } from './modules/auth/auth.module'
+import { FormModule } from './modules/form/form.module'
+import { PrismaModule } from './modules/prisma/prisma.module'
+import { UserModule } from './modules/user/user.module'
 
 @Module({
 	imports: [
