--- backend/src/auth/auth.module.ts
@@ -1,12 +1,0 @@
-import { Module } from '@nestjs/common'
-
-import { UserService } from '@/user/user.service'
-
-import { AuthController } from './auth.controller'
-import { AuthService } from './auth.service'
-
-@Module({
-	controllers: [AuthController],
-	providers: [AuthService, UserService]
-})
-export class AuthModule {}
