+++ backend/src/libs/common/decorators/is-passwords-matching-constraint.decorator.ts
@@ -4,7 +4,7 @@ import {
 	ValidatorConstraintInterface
 } from 'class-validator'
 
-import { RegisterDto } from '@/auth/dto/register.dto'
+import { RegisterDto } from '@/modules/auth/dto/register.dto'
 
 @ValidatorConstraint({ name: 'IsPasswordsMatching', async: false })
 export class IsPasswordsMatchingConstraint
