+++ backend/src/modules/auth/auth.module.ts
@@ -1,0 +1,14 @@
+import { Module } from '@nestjs/common'
+
+import { UserService } from '@/modules/user/user.service'
+
+import { FormService } from '../form/form.service'
+
+import { AuthController } from './auth.controller'
+import { AuthService } from './auth.service'
+
+@Module({
+	controllers: [AuthController],
+	providers: [AuthService, UserService]
+})
+export class AuthModule {}
