+++ backend/src/modules/auth/decorators/roles.decorator.ts
@@ -1,0 +1,5 @@
+import { SetMetadata } from '@nestjs/common'
+
+export const ROLES_KEY = 'roles'
+
+export const Roles = (...roles: string[]) => SetMetadata(ROLES_KEY, roles)
