+++ backend/src/modules/auth/guards/auth.guard.ts
@@ -1,0 +1,30 @@
+import {
+	CanActivate,
+	ExecutionContext,
+	ForbiddenException,
+	Injectable,
+	UnauthorizedException
+} from '@nestjs/common'
+
+import { UserService } from '@/modules/user/user.service'
+
+@Injectable()
+export class AuthGuard implements CanActivate {
+	public constructor(private readonly userService: UserService) {}
+
+	public async canActivate(context: ExecutionContext): Promise<boolean> {
+		const request = context.switchToHttp().getRequest()
+
+		if (typeof request.session.userId === 'undefined') {
+			throw new UnauthorizedException(
+				'Пользователь не авторизован. Войдите в систему для получения доступаы'
+			)
+		}
+
+		const user = await this.userService.findById(request.session.userId)
+
+		request.user = user
+
+		return true
+	}
+}
