+++ backend/src/modules/prisma/prisma.module.ts
@@ -1,0 +1,10 @@
+import { Global, Module } from '@nestjs/common'
+
+import { PrismaService } from './prisma.service'
+
+@Global()
+@Module({
+	providers: [PrismaService],
+	exports: [PrismaService]
+})
+export class PrismaModule {}
