+++ backend/src/modules/user/user.controller.ts
@@ -1,0 +1,26 @@
+import { Controller, Get, HttpCode, HttpStatus } from '@nestjs/common'
+import { UserRole } from '@prisma/__generated__'
+
+import { Authorization } from '../auth/decorators/auth.decorator'
+import { Authorized } from '../auth/decorators/authorized.decorator'
+
+import { UserService } from './user.service'
+
+@Controller('users')
+export class UserController {
+	constructor(private readonly userService: UserService) {}
+
+	@Authorization()
+	@HttpCode(HttpStatus.OK)
+	@Get('profile')
+	public async findProfile(@Authorized('id') userId: string) {
+		return this.userService.findById(userId)
+	}
+
+	@Authorization(UserRole.ADMIN)
+	@HttpCode(HttpStatus.OK)
+	@Get('by-id/:id')
+	public async findById(@Authorized('id') userId: string) {
+		return this.userService.findById(userId)
+	}
+}
