--- backend/src/prisma/prisma.service.ts
@@ -1,16 +1,0 @@
-import { Injectable, OnModuleDestroy, OnModuleInit } from '@nestjs/common'
-import { PrismaClient } from '@prisma/__generated__'
-
-@Injectable()
-export class PrismaService
-	extends PrismaClient
-	implements OnModuleInit, OnModuleDestroy
-{
-	public async onModuleInit(): Promise<void> {
-		await this.$connect()
-	}
-
-	public async onModuleDestroy(): Promise<void> {
-		await this.$disconnect()
-	}
-}
