--- backend/src/user/user.controller.ts
@@ -1,8 +1,0 @@
-import { Controller } from '@nestjs/common'
-
-import { UserService } from './user.service'
-
-@Controller('users')
-export class UserController {
-	constructor(private readonly userService: UserService) {}
-}
