--- backend/src/user/user.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { UserService } from './user.service';
-import { UserController } from './user.controller';
-
-@Module({
-  controllers: [UserController],
-  providers: [UserService],
-})
-export class UserModule {}
